// src/app/models/certificate-models/certificate.model.ts
import internal from "events";

export class CertificateModel {
    id?: string;

    assessmentMethod?: string;

    assessmentScore?: string;

    qualitativeResult?: string;

    maxScore?: internal;

    pathCertificate?: string;

    observation?: string;

    subscriptionId: string;

    assignmentSchool?: string;
}